Add files via upload
Sifts through the fastq output from FLASH2 to identify which sequences fall into a pre-defined library, and which don't. Sequences which match a library but return no oligo, either for a single or duo library, are reported in the library match file with the records preserved in an additional fastq file.